+++ src/moveproof/__init__.py
@@ -1,6 +1,6 @@
 from .compare import compare_snapshots
 from .fingerprint import FileChangedError, fingerprint_file
-from .model import Change, ChangeSet, FileRecord, Snapshot
+from .model import Change, ChangeSet, FileRecord, ScanIssue, Snapshot
 from .snapshot import create_snapshot, load_snapshot, save_snapshot
 
 __all__ = [
@@ -8,6 +8,7 @@ __all__ = [
     "ChangeSet",
     "FileChangedError",
     "FileRecord",
+    "ScanIssue",
     "Snapshot",
     "compare_snapshots",
     "create_snapshot",
@@ -15,4 +16,3 @@ __all__ = [
     "load_snapshot",
     "save_snapshot",
 ]
-